% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fun98}
\alias{fun98}
\title{A Gibbs sampler using Rcpp}
\usage{
fun98(a, b, N)
}
\arguments{
\item{a}{the fixed parameter}

\item{b}{the fixed parameter}

\item{N}{the fixed parameter}
}
\value{
a chain with tagret joint density \code{x}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
a<-1
b<-1
N<-10000
X1<-fun98(a,b,N)
plot(X1[,1],X1[,2],xlab = "x",ylab = "y")
}
}
