% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21056R.R
\name{myfl}
\alias{myfl}
\title{Record the order and frequency of samples in phenotype data}
\usage{
myfl(pheno, dygx)
}
\arguments{
\item{pheno}{the phenotype data}

\item{dygx}{a document about the sample information}
}
\value{
\code{v3}
}
\description{
For longitudinal data analysis, we need transform the SNP data into longitudinal data. So it is necessary to record the information of phenotype data, for example, the order and measurement number of sample. Then we can construct the longitudinal genotype data. There is a document about the relationship of genotype data and phenotype data.
}
\examples{
\dontrun{
v<-myfl(pheno,dygx)
}
}
