% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21056R.R
\name{myqc}
\alias{myqc}
\title{Select genotype variables by Annovar}
\usage{
myqc(df, df2, pp2, cadd, sift)
}
\arguments{
\item{df}{the exonic variant function by gene based annotation}

\item{df2}{the other exonic variant function by gene based annotation}

\item{pp2}{the database of PolyPhen2 scores}

\item{cadd}{the database of cadd scores}

\item{sift}{the database of sift scores}
}
\value{
a  vector of important gene variables \code{Allrs}
}
\description{
The genotype data has high dimension, so it is difficult to analysis by many methods.  
Thus We can have a preliminary screening by ANNOVAR. ANNOVAR is a good way to handle with the SNP
 data, so we can apply the ideas to R. Firstly, import the exonic variant function by gene based 
 annotation and Score alignment library, PolyPhen2, cadd, sift. The three scores can  assess the 
 importance of genes well. When there are two scores show that the gene is important, then keep it.
}
\examples{
\dontrun{
index<-myqc(df,df2,pp2,cadd,sift)
}
}
