\name{StatComp21056-package}
\alias{StatComp21056-package}
\alias{StatComp21056}
\docType{package}
\title{
   Select gene variables by Annovar
}
\description{
  Several R and Rcpp function are included in to select variables.
}

\author{
Yu Zhang

Maintainer: Yu Zhang <sa21204190zy@mail.ustc.edu.cn>
}
\keyword{ package }
